Firefly-179: response to review and testing
